--- static/js/apps/nl_interface/nl_constants.ts
@@ -1,30 +1,0 @@
-/**
- * Copyright 2022 Google LLC
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-/**
- * Constants and functions shared across components.
- */
-
-export const NL_URL_PARAMS = {
-  IDX: "idx",
-  LLM: "llm",
-  SMALL: "small",
-  MEDIUM: "medium",
-};
-
-export const MAX_QUERY_COUNT = 10;
-
-export const SVG_CHART_HEIGHT = 160;
