+++ static/js/apps/nl_interface/query_history.tsx
@@ -22,7 +22,7 @@ import axios from "axios";
 import _ from "lodash";
 import React, { useEffect, useState } from "react";
 
-import { MAX_QUERY_COUNT } from "./nl_constants";
+import { MAX_QUERY_COUNT } from "../../constants/app/nl_interface_constants";
 
 interface QueryHistoryProps {
   // Callback function for history item clicks.
